laboratorio8/apilab/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
asgi-correlation-id